+++ src/data/usecases/authentication/db-authentication-protocols.ts
@@ -1,6 +1,6 @@
 export * from '../../../domain/usecases/authentication'
-export * from '../../protocols/db/load-account-by-email-repository'
 export * from '../add-account/db-add-account-protocols'
+export * from '../../protocols/db/account/load-account-by-email-repository'
 export * from '../../protocols/criptography/hash-comparer'
 export * from '../../protocols/criptography/encrypter'
-export * from '../../protocols/db/update-access-token-repository'
+export * from '../../protocols/db/account/update-access-token-repository'
