+++ src/presentation/controllers/signup/signup-controller.ts
@@ -1,4 +1,4 @@
-import { HttpResponse, HttpRequest, Controller, AddAccount } from './signup-protocols'
+import { HttpResponse, HttpRequest, Controller, AddAccount } from './signup-controller-protocols'
 import { badRequest, ok, serverError } from '../../helpers/http/http-helper'
 import { Validation } from '../../protocols/validation'
